Move proxy::Server to Proxy and sessions::session to sessions
