+++ docs/src/filters/firewall.md
@@ -34,7 +34,7 @@ clusters:
 # ";
 # let config = quilkin::config::Config::from_reader(yaml.as_bytes()).unwrap();
 # assert_eq!(config.filters.load().len(), 1);
-# quilkin::Server::try_from(config).unwrap();
+# quilkin::Proxy::try_from(config).unwrap();
 ```
 
 ### Configuration Options ([Rust Doc](../../api/quilkin/filters/firewall/struct.Config.html))
