+++ docs/src/filters/load_balancer.md
@@ -25,7 +25,7 @@ clusters:
 # ";
 #   let config = quilkin::config::Config::from_reader(yaml.as_bytes()).unwrap();
 # assert_eq!(config.filters.load().len(), 1);
-# quilkin::Server::try_from(config).unwrap();
+# quilkin::Proxy::try_from(config).unwrap();
 # }
 ```
