+++ src/lib.rs
@@ -37,7 +37,7 @@ pub mod xds;
 pub type Result<T, E = eyre::Error> = std::result::Result<T, E>;
 
 #[doc(inline)]
-pub use self::{cli::Cli, config::Config, proxy::Server};
+pub use self::{cli::Cli, config::Config, proxy::Proxy};
 
 pub use quilkin_macros::include_proto;
