+++ tests/capture.rs
@@ -33,7 +33,7 @@ async fn token_router() {
     let mut t = TestHelper::default();
     let echo = t.run_echo_server().await;
     let server_port = 12348;
-    let server_config = quilkin::Server::builder()
+    let server_config = quilkin::Config::builder()
         .port(server_port)
         .filters(vec![
             Filter {
