+++ tests/filter_order.rs
@@ -55,7 +55,7 @@ on_write: DECOMPRESS
         .await;
 
     let server_port = 12346;
-    let server_config = quilkin::Server::builder()
+    let server_config = quilkin::Config::builder()
         .port(server_port)
         .filters(vec![
             Filter {
