+++ tests/load_balancer.rs
@@ -47,7 +47,7 @@ policy: ROUND_ROBIN
     }
 
     let server_port = 12346;
-    let server_config = quilkin::Server::builder()
+    let server_config = quilkin::Config::builder()
         .port(server_port)
         .filters(vec![Filter {
             name: LoadBalancer::factory().name().into(),
