+++ tests/token_router.rs
@@ -44,7 +44,7 @@ quilkin.dev:
         - YWJj # abc
         ";
     let server_port = 12348;
-    let server_config = quilkin::Server::builder()
+    let server_config = quilkin::Config::builder()
         .port(server_port)
         .filters(vec![
             Filter {
